Modificacoes em classes MotoristaRepository, MotoristaResource,MotoristaService e BDService

+++ src/main/java/com/vamos/repositories/MotoristaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.vamos.models.Motorista;
 
 @Repository
-public interface MotoristaRepository extends JpaRepository<Motorista, Integer> {
+public interface MotoristaRepository extends JpaRepository<Motorista, Long> {
 
 }
